feat: add Astro example shoe store
A new `apps/astro` app has been added. I've tried to keep it as simple as
possible, and as close to the other shoe stores as I could. This means that the
markup is copy/pasted with only very few changes (removal of `key` attributes,
renaming `className` to `class`, etc). Some of the markup is rendered
conditionally based on a loading flags, and even though these will always
evaluate to `false` for now, I decided to keep them. When we add visual
editing, we'll begin to need them. But for now, the pages are rendered using
SSR with no need for loading states.