fix(presentation): remove use of startTransition
When investigating #1486 inside
https://github.com/sanity-io/template-nextjs-personal-website I found that the
`liveQueries` in `PresentationTool` easily run into an infinite loop causing
the component to crash due to `Maximum update depth exceeded`. When inspecting
the changes to the `liveQueries` that caused the re-renders I didn't find any
differences expect a slight difference in `receivedAt`. Proceeding with the
investigation, I couldn't find any odd places where `setLiveQueries` was called
or where the `liveQueries` might be mutated. However, as soon as I remove the
use of `startTransition`, when expiring the queries at a regular interval, the
problem goes away. The component, `PresentationTool`, suddenly feels completely
stable and nothing I do in the UI can provoke the infinite loop again. I'm not,
however, sure how `startTransition` could cause this issue or what implications
it might have that we remove it. All I know is that it fixes the issue at hand.