feat(astro): add visual editing POC
It's not much, but I think it's enough to get the ball rolling on visual
editing in Astro.

All queries are now passed through a `loadQuery` function similar to the one
found in the Next.js server-only example. Here, we set the proper config to
support visual editing in case that env var is set to `'true'`. Similarly, a
`<VisualEditing />` component is added to the root layout in order to toggle
visual editing if enabled. The `<VisualEditing />` component just triggers a
full page refresh whenever a refresh should happen.

Some questions that pop up are:

- What is the best way to toggle visual editing in Astro? Is it an env var like
  we use here, or is there a better way a la `draftMode` in Next.js?
- How do you best encapsulate the `<VisualEditing />` component to make sure
  its script is only loaded and run when visual editing is enabled? In the
  current implementation, `enableVisualEditing` is only called if the env var
  flag is set, but the `@sanity/visual-editing` module is still included, which
  might not be ideal. I've tried different techniques outlined in
  https://docs.astro.build/en/guides/client-side-scripts/, and the
  data-attribute-passed-to-custom-component is the technique I found to work
  the best. But maybe I missed something.
- Is there a better way to refresh the page on changes than using
  `window.location.reload()`?